LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-position-anchor-expected.txt


FAIL #anchored 1 assert_equals:
<div id="anchored" data-offset-x="100"></div>
offsetLeft expected 100 but got 0
